6.0.0.1，代码重构，通讯协议升级，中间服务重构(未完成)

+++ SiMay.Net.SessionProvider/SessionProviderOptions.cs
@@ -9,12 +9,24 @@ namespace SiMay.Net.SessionProvider
 {
     public class SessionProviderOptions
     {
+        /// <summary>
+        /// Session提供类型
+        /// </summary>
         public SessionProviderType SessionProviderType { get; set; }
 
-        public int AccessId { get; set; }
+        /// <summary>
+        /// 中间服务识别的唯一ID
+        /// </summary>
+        public long AccessId { get; set; }
 
+        /// <summary>
+        /// 中间服务访问Key
+        /// </summary>
         public long AccessKey { get; set; }
 
+        /// <summary>
+        /// 中间服务地址
+        /// </summary>
         public IPEndPoint ServiceIPEndPoint { get; set; }
 
         public int PendingConnectionBacklog { get; set; }
